+++ src/main/java/pl/edu/pw/elka/paprykaisalami/geeruh/spring/config/GeneralConfig.java
