+++ src/main/java/pl/edu/pw/elka/paprykaisalami/geeruh/spring/errors/ErrorsException.java
@@ -1,4 +1,4 @@
-package pl.edu.pw.elka.paprykaisalami.geeruh.errors;
+package pl.edu.pw.elka.paprykaisalami.geeruh.spring.errors;
 
 import lombok.Builder;
 import lombok.Getter;
